assorted updates and fixes

+++ bin/bin/backup.sh
@@ -2,22 +2,35 @@
 
 #DEFINTIONS!
 DESTINATION=/media/Toshiba_Backups
-
 list_files() {
   \ls -a | grep -v \^$USER\$ | grep -v \^.\$ | grep -v \^..\$
 }
 
-# Nothing should be hardcoded beyond ~/.dotfiles
-if [ -d "$DESTINATION/" ]; then
-  cd $DESTINATION
-  mkdir $USER
-  mv `list_files` $USER/ 
-  rsync -aAv --delete --exclude=.cache $HOME $DESTINATION/
-  rsync -aAv --delete $HOME/.dotfiles/* $DESTINATION/$USER/.dotfiles/
-  mv $DESTINATION/$USER/* $DESTINATION
-  cd $USER
-  mv `list_files` $DESTINATION
-  rmdir $DESTINATION/$USER
-  exit $?
-else echo $DESTINATION is not found ; exit 1
+#Check user
+
+if  [ -f $DESTINATION/$USER ]; then
+
+# No files/directories should be hardcoded
+  if [ -d "$DESTINATION/" ]; then
+    chmod -R +w $DESTINATION
+    cd $DESTINATION
+    rm $DESTINATION/$USER
+    #ls #debug
+    #file $USER  #debug
+    #sleep 30 #debug
+    mkdir $USER
+    mv `list_files` $USER/ 
+    rsync -aAv --delete --exclude=.cache $HOME $DESTINATION/
+    mv $DESTINATION/$USER/* $DESTINATION
+    cd $USER
+    mv `list_files` $DESTINATION
+    rmdir $DESTINATION/$USER
+    #sleep 40 #debug
+    echo $USER > $DESTINATION/$USER
+    chmod -R a-w $DESTINATION
+    exit $?
+  else echo $DESTINATION is not found ; exit 1
+  fi
+
+else echo $DESTINATION not authorized for write by current user ; exit 10
 fi
